qcacmn: Modify data type of timestamp
The 'timestamp' field in CFR meta data corresponds to TSF
at the reception of PPDU in micro second unit. To capture
timestamp for one shot, periodic and repetitive channel
capture, this field is being used.

case 1:
For one shot and periodic capture, the clock used for
timestamp is private to the target and not visible to host.
So it is being passed by FW to host and the placeholder
being exchanged between Host and target is of 32 bit.

case 2:
In case of repetitive channel capture, the
timestamp is being populated from the received PPDU,
which is 64bit.

To accommodate the timestamp, in both above cases, change
the data type to u_int64_t from u_int32_t

CRs-Fixed: 2751408

+++ umac/cfr/dispatcher/inc/wlan_cfr_utils_api.h
@@ -177,7 +177,7 @@ struct cfr_metadata_version_3 {
 	u_int8_t    capture_type; /* cfr_capture_type */
 	u_int8_t    sts_count;
 	u_int8_t    num_rx_chain;
-	u_int32_t   timestamp;
+	u_int64_t   timestamp;
 	u_int32_t   length;
 	u_int8_t    is_mu_ppdu;
 	u_int8_t    num_mu_users;
